(fix): store an inline double with size 0 like the float endpoint
Attribute::SIZES fixes the size of the types that do not take one, and
main keyed the double entry on Database::VAR_FLOAT, which is 'double'.
The enum migration keyed it on ColumnType::Float, 'float', a type the
inline validator rejects, so the entry could never match and an inline
{"type": "double", "size": 5} on create collection/table stored size 5
where createFloatColumn and main store 0. The column is a DOUBLE either
way; the stored size is what list attributes/columns filters on and what
migrations compare against an existing attribute.

The entry is keyed on ColumnType::Double again. The two resolve()
assertions in AttributesTest that the migration moved from VAR_FLOAT to
ColumnType::Float go back to the double they pinned on main.

+++ src/Appwrite/Utopia/Database/Attribute.php
@@ -27,7 +27,7 @@ class Attribute
         // Every adapter maps these two types without consulting the size.
         ColumnType::BigInteger->value => 8,
         'bigint' => 8,
-        ColumnType::Float->value => 0,
+        ColumnType::Double->value => 0,
     ];
 
     /**
